+++ src/main/java/nl/nazeem/babel/BabelServiceApplication.java
@@ -1,4 +1,4 @@
-package nl.nazeem.service.babel;
+package nl.nazeem.babel;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
